[BE] Feat: 학생/강사 수업 분석 결과 API

See merge request s09-webmobile1-sub2/S09P12B105!183

+++ BackEnd/src/main/java/com/learnershigh/dto/etc/AttentionDto.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class AttentionDto {
 
     public Id id;
-    public float avgValue;
+    public Double avgValue;
     public int count0;
     public int count1;
     public int count2;
